requirements.txt
django >= 3.1
pandas
django-import-export
psycopg2
nltk
# additionally execute nltk.download() to download corpora
six
django-extensions
django-cors-headers
djangorestframework
django-language-fields
django-model-utils
markdown
django-filter
requests
# scheduled task
apscheduler
requests-toolbelt
django-apscheduler

# BERT-recommender
tensorflow >=2.3
transformers ==3.1
cloud-tpu-client

#model-downloader
#how to install seafile-cli on your system: see https://download.seafile.com/published/seafile-user-manual/syncing_client/install_linux_client.md
git+https://github.com/cstenkamp/python-seafile.git@v0.1.2#egg=python_seafile #on pypi there's only v0.1.0, which is broken, and even the original repo has an error with binary files

#dashboard
plotly >= 4.14
matplotlib
pydot
sickle
vobject
